Split view and controller into separate scripts
